==> 01-Data_Cleaning.Rmd <==
---
title: "Step I: Data Cleaning"
subtitle: "PACs and Perceptions"
author:
  - name: Nicholas R. Jenkins 
    affiliation: University of California, Riverside
    affiliation_url: https://www.ucr.edu
abstract: |
  This document shows all the code used to clean the data for this project. 
  Each step has been carefully documented and will replicate the data set used
  in the study. For any questions about this code, please contact 
  Nick Jenkins at [nicholas.jenkins@email.ucr.edu](mailto:nicholas.jenkins@email.ucr.edu). 
  All data cleaning was done using RStudio running R version `r R.version$major`
  .`r R.version$minor` "`r R.version$nickname`".
date: "`r Sys.Date()`"
output: 
  html_notebook: 
    highlight: pygments
    theme: paper
    toc: yes
    code_folding: 
    toc_float: true
bibliography: /Users/nick/Documents/Research/References/BibTeX/biblatex.bib
editor_options: 
  chunk_output_type: inline
---

```{r setup, include = FALSE}
# Clear Environment -----------------------------------------------------------
rm(list = ls())

# Load Packages ---------------------------------------------------------------
packages <- c("tidyverse", "stargazer", "foreign", "sqldf", "readxl", 
              "openintro", "lubridate")
lapply(packages, require, character.only = TRUE)

# Inline Formatting -----------------------------------------------------------
comma <- function(x) format(x, digits = 2, big.mark = ",")

# Set Global Chunk Options ----------------------------------------------------
knitr::opts_chunk$set(
  echo = TRUE,
  warning = FALSE,
  message = FALSE,
  comment = "##",
  R.options = list(width = 70)
)
```


# Import Omnibus Data

```{r}
data.raw <- read_csv("Data/Raw Data/omnibusS20_data.csv")

estimation.data <- 
  data.raw %>%
  rename(vote = Q144,
         donate = Q145,
         trustworthy = Q146,
         pac_treat = m6_cont,
         experience = m6_prior,
         race = m6_racenat,
         gender = Q4) %>%
  select(vote, donate, trustworthy, party_id, pac_treat, experience, race,
         party_id, gender)

estimation.data <- 
  estimation.data %>%
  mutate(vote = factor(vote, 
                       levels = c(1:7), 
                       labels = c("Very Unlikely", "Unlikely", 
                                  "Slightly Unlikely", 
                                  "Neither Unlikely nor Likely",
                                  "Slightly Likely", "Likely", "Very Likely"),
                       ordered = TRUE),
         donate = factor(donate,
                         levels = c(1:7),
                         labels = c("Very Unlikely", "Unlikely", 
                                  "Slightly Unlikely", 
                                  "Neither Unlikely nor Likely",
                                  "Slightly Likely", "Likely", "Very Likely"),
                         ordered = TRUE),
         trustworthy = factor(trustworthy, 
                              levels = c(1:7),
                              labels = c("Very Untrustworthy", "Untrustworthy", 
                                  "Slightly Untrustworthy", 
                                  "Neither Untrustworthy nor Trustworthy",
                                  "Slightly Trustworthy", "Trustworthy", 
                                  "Very Trustworthy"),
                              ordered = TRUE),
         pac_treat = factor(pac_treat,
                            levels = c(1:2),
                            labels = c("Does Not Accept PAC Contributions",
                                       "Accepts PAC Contributions")),
         experience = factor(experience,
                             levels = c(1:3),
                             labels = c("No Prior Experience", 
                                        "Councilmember", "Mayor")),
         race = factor(race, 
                       levels = c(1:5),
                       labels = c("White", "Mexican American", 
                                  "Salvadoran American", "Chinese American",
                                  "Vietnamese American")),
         party_id = as_factor(party_id),
         gender = factor(gender, 
                         levels = c(1, 2, 3), 
                         labels = c("Famle", "Male", "Nonbinary")))
```


# Final Data Set Summary Statistics 

```{r Summary Statistics, echo = TRUE}
summary(estimation.data)

save(estimation.data, file = "Data/Clean Data/estimation_data.Rda")
```
